Projekt-Automatisierung-Kunden/requirements.txt
openai>=1.14.0
python-dotenv
requests
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytz

